requirements.txt
spotipy
python-dotenv
openai
psycopg[binary]
SQLAlchemy
python-dotenv

